handle window dimension change from Router via redux store
