+++ app/src/index.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import { View, Dimensions } from "react-native";
+import { View } from "react-native";
 import { Provider } from "react-redux";
 import store from "./store";
 
@@ -8,23 +8,8 @@ import { imageAssets } from "theme/images";
 import { fontAssets } from "theme/fonts";
 import Router from "./components/routes";
 
-// Dimensions.
-const window = Dimensions.get("window");
-
 const App = () => {
   const [didLoad, setDidLoad] = useState(false);
-  const [dimensions, setDimensions] = useState({ window });
-  
-  const onChange = ({ window }) => {
-    setDimensions({ window });  
-  };
-  
-  useEffect(() => {
-      Dimensions.addEventListener("change", onChange);
-      return () => {
-        Dimensions.removeEventListener("change", onChange);  
-      };
-  }, []);
 
   const handleLoadAssets = async () => {
     // Asset preloading.
